fix some bug

+++ system/dosen.php
@@ -241,9 +241,12 @@ class Dosen
         if (isset($inputPhoto) && $inputPhoto['error'] == 0) {
 
             $file = getImageUpload($_POST['user_id'], 'Users.users');
-            $filepath = "../" . $file['data'][0]['path'];
-           if ($file['num_rows'] > 0 && file_exists($filepath)) {
-                unlink($filepath);
+            
+            if($file['num_rows'] > 0) {
+                $filepath = "../" . $file['data'][0]['path'];
+                if ($file['num_rows'] > 0 && file_exists($filepath)) {
+                    unlink($filepath);
+                }
             }
 
             $fileType = strtolower(pathinfo($inputPhoto["name"], PATHINFO_EXTENSION));
